requirements.txt
joblib==1.1.0
numpy==1.22.4
pytest==7.1.2
requests==2.27.1
scikit-learn==1.1.1
scipy==1.8.1
threadpoolctl==3.1.0
uvicorn==0.32.1
PyYAML==6.0.2
flask==3.1.0
fastapi==0.115.6
httpx==0.28.1
prometheus_client==0.21.1